Adds tests for s3 and divshot deployment
This adds deployments for S3 and Divshot, should be a little easier to
create tests in the future now.

This also creates a new JavaScript API module, deploy.js with a
signature of:

```js
deploy(packageJSON, deployConfig, selectedConfig, files, errback)
```

This just pulled out stuff that was in bin/deploy, making that module
much smaller.